Move getPlugins to plugins.go
